+++ src/action/input.rs
@@ -1,7 +1,6 @@
 use crate::instant::Instant;
 
 /// The input of [`Action::exec`](crate::action::Action::exec).
-/// `self.instant().rest()` is guaranteed to be non-empty.
 ///
 /// Once created, only [`Self::state`] and [`Self::heap`] can be mutated.
 ///
@@ -37,51 +36,19 @@ pub struct Input<'text, StateRef, HeapRef> {
 }
 
 impl<'text, StateRef, HeapRef> Input<'text, StateRef, HeapRef> {
-  /// # Safety
-  /// You should ensure that [`Instant::rest`] is not empty.
-  /// This will be checked using [`debug_assert!`].
-  /// For the checked version, see [`Self::new`].
   #[inline]
-  pub const unsafe fn new_unchecked(
-    instant: Instant<'text>,
-    state: StateRef,
-    heap: HeapRef,
-  ) -> Self {
-    debug_assert!(!instant.rest().is_empty());
-    Self {
-      instant,
+  pub const fn new(instant: Instant<'text>, state: StateRef, heap: HeapRef) -> Self {
+    Input {
       state,
       heap,
+      instant,
     }
   }
 
-  /// Return [`Some`] if [`Instant::rest`] is not empty.
-  #[inline]
-  pub fn new(instant: Instant<'text>, state: StateRef, heap: HeapRef) -> Option<Self> {
-    (!instant.rest().is_empty()).then(|| unsafe { Self::new_unchecked(instant, state, heap) })
-  }
-
   /// The [`Instant`] before this action is executed.
-  /// [`Instant::rest`] is guaranteed to be non-empty.
   #[inline]
   pub const fn instant(&self) -> &Instant<'text> {
     &self.instant
-  }
-
-  /// The first char in [`Instant::rest`].
-  ///
-  /// Since `self.instant().rest()` is guaranteed to be non-empty,
-  /// the next char is guaranteed to exist.
-  ///
-  /// This is faster than `self.instant().rest().chars().next().unwrap()`.
-  ///
-  /// This value is not stored in this struct
-  /// because the value is not always needed.
-  /// You can cache the return value as needed.
-  #[inline]
-  pub fn next(&self) -> char {
-    // SAFETY: `self.instant.rest()` is guaranteed to be not empty.
-    unsafe { self.instant.rest().chars().next().unwrap_unchecked() }
   }
 }
 
@@ -99,74 +66,50 @@ impl<'text, State, Heap> Input<'text, &mut State, &mut Heap> {
     }
   }
 }
+
+// TODO: make this a function?
 /// Construct a new [`Input`] by digesting `n` bytes from [`Input::instant`].
-///
-/// Return [`Some`] if [`Instant::rest`] of the new instant is not empty.
 /// # Safety
 /// You should ensure that `n` is a valid UTF-8 boundary.
 /// This will be checked using [`debug_assert!`].
 /// # Performance
 /// This is a macro to make sure this is always inlined.
 macro_rules! shift_input {
-  ($input:expr, $n:expr) => {
+  ($input:expr, $n:expr) => {{
     // perf: check the len first to prevent unnecessary clone of instant
-    ($n < $input.instant().rest().len()).then(|| {
-      let mut instant = $input.instant().clone();
-      unsafe {
-        instant.digest_unchecked($n);
-        Input::new_unchecked(instant, &mut *$input.state, &mut *$input.heap)
-      }
-    })
-  };
+    let mut instant = $input.instant().clone();
+    unsafe { instant.digest_unchecked($n) };
+    Input::new(instant, &mut *$input.state, &mut *$input.heap)
+  }};
 }
 pub(crate) use shift_input;
-
 #[cfg(test)]
 mod tests {
   use super::*;
 
   #[test]
-  fn input_new_unchecked() {
-    let mut state = ();
-    let mut heap = ();
-    let input = unsafe { Input::new_unchecked(Instant::new("123"), &mut state, &mut heap) };
-    assert_eq!(input.instant().text(), "123");
-    assert_eq!(input.instant().rest(), "123");
-    assert_eq!(input.instant().digested(), 0);
-    assert_eq!(input.next(), '1');
-  }
-
-  #[test]
-  #[should_panic]
-  fn input_new_unchecked_empty() {
-    unsafe { Input::new_unchecked(Instant::new(""), &mut (), &mut ()) };
-  }
-
-  #[test]
   fn input_new() {
     let mut state = ();
     let mut heap = ();
-    let input = Input::new(Instant::new("123"), &mut state, &mut heap).unwrap();
+    let input = Input::new(Instant::new("123"), &mut state, &mut heap);
     assert_eq!(input.instant().digested(), 0);
     assert_eq!(input.instant().rest(), "123");
-    assert_eq!(input.next(), '1');
   }
 
   #[test]
-  fn input_new_no_rest() {
-    assert!(Input::new(Instant::new(""), &mut (), &mut ()).is_none());
+  fn input_new_no_rest_is_ok() {
+    Input::new(Instant::new(""), &mut (), &mut ());
   }
 
   #[test]
   fn input_reborrow() {
     let mut state = 123;
     let mut heap = 123;
-    let mut input = Input::new(Instant::new("123"), &mut state, &mut heap).unwrap();
+    let mut input = Input::new(Instant::new("123"), &mut state, &mut heap);
     {
       let input = input.reborrow();
       assert_eq!(input.instant().digested(), 0);
       assert_eq!(input.instant().rest(), "123");
-      assert_eq!(input.next(), '1');
       *input.state = 456;
       *input.heap = 456;
     }