+++ src/combinator/decorator/state.rs
@@ -186,7 +186,7 @@ mod tests {
       .prepare(|input| {
         input.state.from = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_some());
     assert_eq!(state, State { from: 1, to: 1 });
   }
@@ -198,7 +198,7 @@ mod tests {
       .then(|ctx| {
         ctx.input.state.from = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_some());
     assert_eq!(state, State { from: 1, to: 0 });
 
@@ -207,7 +207,7 @@ mod tests {
       .then(|ctx| {
         ctx.input.state.from = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_none());
     assert_eq!(state, State { from: 0, to: 0 });
   }
@@ -219,7 +219,7 @@ mod tests {
       .catch(|input| {
         input.state.from = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_some());
     assert_eq!(state, State { from: 0, to: 0 });
 
@@ -228,7 +228,7 @@ mod tests {
       .catch(|input| {
         input.state.from = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_none());
     assert_eq!(state, State { from: 1, to: 0 });
   }
@@ -240,7 +240,7 @@ mod tests {
       .finally(|input| {
         input.state.to = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_some());
     assert_eq!(state, State { from: 0, to: 1 });
 
@@ -249,7 +249,7 @@ mod tests {
       .finally(|input| {
         input.state.to = 1;
       })
-      .exec(Input::new(Instant::new("123"), &mut state, &mut ()).unwrap())
+      .exec(Input::new(Instant::new("123"), &mut state, &mut ()))
       .is_none());
     assert_eq!(state, State { from: 0, to: 1 });
   }